+++ src/app/modules/command/service/command.service.ts
@@ -11,10 +11,12 @@ export class CommandService {
         private readonly keyboard: KeyboardService) { }
 
     public command(command: string): void {
+        const text = this.clipboard.readText();
         this.keyboard.setKeyboardDelay(5);
         this.keyboard.keyTap('enter');
         this.clipboard.writeText(command);
         this.keyboard.keyTap('v', ['control']);
         this.keyboard.keyTap('enter');
+        this.clipboard.writeText(text);
     }
 }