+++ src/app/modules/evaluate/component/evaluate-dialog/evaluate-dialog.component.ts
@@ -7,11 +7,11 @@ import { CurrencyService } from '@shared/module/poe/service/currency/currency.se
 import { Item, ItemRarity, ItemSearchEvaluateResult, Language } from '@shared/module/poe/type';
 import { BehaviorSubject, forkJoin, Observable, of, Subject } from 'rxjs';
 import { debounceTime, flatMap, switchMap, takeUntil } from 'rxjs/operators';
+import { EvaluateUserSettings } from '../evaluate-settings/evaluate-settings.component';
 
 export interface EvaluateDialogData {
   item: Item;
-  currencyId: string;
-  queryDefault: boolean;
+  settings: EvaluateUserSettings;
   language?: Language;
 }
 
@@ -61,6 +61,7 @@ export class EvaluateDialogComponent implements OnInit {
       category: item.category,
       rarity: item.rarity,
       corrupted: item.corrupted,
+      veiled: item.veiled,
       influences: item.influences || {},
       damage: {},
       stats: [],
@@ -69,7 +70,7 @@ export class EvaluateDialogComponent implements OnInit {
       sockets: [],
     };
 
-    if (this.data.queryDefault) {
+    if (this.data.settings.evaluateQueryDefault) {
       if (item.level && item.level > 80) {
         queryItem.level = item.level;
       }
@@ -107,9 +108,13 @@ export class EvaluateDialogComponent implements OnInit {
   }
 
   private search(item: Item): Observable<ItemSearchEvaluateResult> {
+    const currencies$ = this.data.settings.evaluateCurrencyIds.map(id => this.currencyService.searchById(id));
     return forkJoin(
-      this.itemSearchService.search(item),
-      this.currencyService.searchById(this.data.currencyId)
+      this.itemSearchService.search(item, {
+        indexed: this.data.settings.evaluateQueryIndexed,
+        online: this.data.settings.evaluateQueryOnline,
+      }),
+      forkJoin(currencies$)
     ).pipe(
       flatMap(results => {
         if (results[0].items.length <= 0) {