+++ src/app/modules/tool/service/tool.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { KeyboardService } from '@app/service';
+import { KeyboardService, MouseService, WindowService } from '@app/service';
 import { BehaviorSubject } from 'rxjs';
 import { throttleTime } from 'rxjs/operators';
 
@@ -10,7 +10,9 @@ export class ToolService {
     private readonly command$ = new BehaviorSubject<string>('');
 
     constructor(
-        private readonly keyboard: KeyboardService) {
+        private readonly keyboard: KeyboardService,
+        private readonly mouse: MouseService,
+        private readonly window: WindowService) {
         this.init();
     }
 
@@ -20,18 +22,25 @@ export class ToolService {
 
     private init(): void {
         this.command$.pipe(
-            throttleTime(25)
+            throttleTime(15)
         ).subscribe(command => {
-            this.keyboard.setKeyboardDelay(5);
-            switch (command) {
-                case 'storage-left':
-                    this.keyboard.keyTap('left');
-                    break;
-                case 'storage-right':
-                    this.keyboard.keyTap('right');
-                    break;
-                default:
-                    break;
+            const gameBounds = this.window.getBounds();
+            const stashWidth = gameBounds.width / 2.88;
+
+            const point = this.mouse.getCursorScreenPoint();
+            const relativePointX = point.x - gameBounds.x;
+            if (relativePointX > stashWidth) {
+                this.keyboard.setKeyboardDelay(5);
+                switch (command) {
+                    case 'storage-left':
+                        this.keyboard.keyTap('left');
+                        break;
+                    case 'storage-right':
+                        this.keyboard.keyTap('right');
+                        break;
+                    default:
+                        break;
+                }
             }
         });
     }